backend/hf_extractor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytesseract==0.3.10
pillow==10.2.0

